issue #288 : fix random BigDecimal generation
Random BigDecimal generation  was based on long instead of double.
When specifying a min as 0.01, a min of 0 was used.
This commit fixes the issue by using the double type to avoid any data
precision loss.

+++ random-beans-validation/src/main/java/io/github/benas/randombeans/validation/DecimalMinMaxAnnotationHandler.java
@@ -101,15 +101,15 @@ class DecimalMinMaxAnnotationHandler implements BeanValidationAnnotationHandler 
             }
             if (fieldType.equals(BigDecimal.class)) {
                 return new BigDecimalRangeRandomizer(
-                        minValue == null ? null : minValue.longValue(),
-                        maxValue == null ? null : maxValue.longValue(),
+                        minValue == null ? null : minValue.doubleValue(),
+                        maxValue == null ? null : maxValue.doubleValue(),
                         random.nextLong()
                 );
             }
             if (fieldType.equals(String.class)) {
                 BigDecimalRangeRandomizer delegate = new BigDecimalRangeRandomizer(
-                        minValue == null ? null : minValue.longValue(),
-                        maxValue == null ? null : maxValue.longValue(),
+                        minValue == null ? null : minValue.doubleValue(),
+                        maxValue == null ? null : maxValue.doubleValue(),
                         random.nextLong()
                 );
                 return new StringDelegatingRandomizer(delegate);
